refactor(dapp): centralize approval methods and shared encoders
