+++ src/lib/dapp/responses.ts
@@ -3,10 +3,11 @@ import type {
   DappRpcResponse,
   DappRuntimePendingAck,
 } from '@/lib/dapp/protocol'
+import { CONTENT_SOURCE, DAPP_CHANNEL } from '@/lib/dapp/protocol'
 
 export const asDappSuccess = (id: string, result: unknown): DappRpcResponse => ({
-  channel: 'qubic:dapp',
-  source: 'qubic:content',
+  channel: DAPP_CHANNEL,
+  source: CONTENT_SOURCE,
   id,
   ok: true,
   result,
@@ -17,17 +18,11 @@ export const asDappFailure = (
   code: DappProviderErrorCode,
   message: string,
 ): DappRpcResponse => ({
-  channel: 'qubic:dapp',
-  source: 'qubic:content',
+  channel: DAPP_CHANNEL,
+  source: CONTENT_SOURCE,
   id,
   ok: false,
   error: { code, message },
 })
 
 export const asRuntimePendingAck = (id: string): DappRuntimePendingAck => ({ pending: true, id })
-
-export const isRuntimePendingAck = (value: unknown): value is DappRuntimePendingAck => {
-  if (!value || typeof value !== 'object') return false
-  const record = value as Record<string, unknown>
-  return record.pending === true && typeof record.id === 'string' && Boolean(record.id)
-}
